pipeline: return to single file
