--- astropop/pipelines/boss.py
@@ -1,7 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-
-import abc
-
-
-class Boss(abc.ABC):
-    """Class to handle the general pipeline management."""
